fix: MultiColumnTfidfVectorizer returns sparse matrix of shape (n_rows, 1) when all words are pruned
Currently when all terms are pruned due to the `max_df` and `min_df` thresholds, a matrix of shape (n_rows, 0) is returned. This sparse matrix will fail if fed into the sklearn `check_array` function. In order to be compatible with this function that is used in `sklearn.compose.ColumnTransformer`, `MultiColumnTfidfVectorizer` has been changed to return a sparse matrix of shape (n_rows, 1) in this situation.